Testing draw code
This code has draw uncommented and set up properly. I am in the process of getting everything to draw correctly, and this code opens a window that when in debug mode shows that everything is drawn in the window (stacked on top of each other), but after 3 loops in the while key pressed event loop towards the bottom of main the window closes automatically and the console stops. I will be working on this but anyone who wants to try and figure it out can go for it.